refactor(core): Unify custom field interceptor processing
